create family member clear

+++ database/migrations/2025_08_28_161411_create_family_members_table.php
@@ -17,9 +17,8 @@ return new class extends Migration
             $table->foreign('head_of_family_id')->references('id')->on('head_of_families');
             $table->uuid('user_id');
             $table->foreign('user_id')->references('id')->on('users');
-
             $table->string('profile_picture');
-            $table->integer('identity_number');
+            $table->bigInteger('identity_number');
             $table->enum('gender', ['male', 'female', 'other']);
             $table->date('date_of_birth');
             $table->string('phone_number');
